# docs/swagger.yaml
basePath: /
definitions:
...
  p2-ip-kartikajls_internal_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  p2-ip-kartikajls_internal_dto.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  p2-ip-kartikajls_internal_helper.Response:
    properties:
      responseCode:
        type: integer
      responseData: {}
      responseMessage:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API untuk aplikasi Rent Camera.
    API ini menyediakan fitur register, login, camera, top up, rental order, dan payment.
    Admin dapat melakukan approval terhadap top up, order, dan payment.
  title: Jati Rent Camera API
  version: "1.0"
paths:
...
  /users/{id}:
    get:
      description: Mendapatkan data user berdasarkan ID
      parameters:
      - description: User ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Get User By ID
      tags:
      - User
...
  /users/login:
    post:
      consumes:
      - application/json
      description: User melakukan login untuk mendapatkan JWT
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/p2-ip-kartikajls_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      summary: Login User
      tags:
      - User - Login
...
  /users/register:
    post:
      consumes:
      - application/json
      description: User melakukan registrasi akun
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/p2-ip-kartikajls_internal_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      summary: Register User
      tags:
      - User - Register
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan JWT dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"